+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - update to 2021 edition
 - update MSRV to 1.72.0
 
-
+- **breaking**: `dotenvy::Error` now includes IO file path info and variable name info
 - **breaking**: `dotenvy::var` no longer calls `load` internally
 - **breaking**: `dotenvy::Result` is now private
 - **breaking**: deprecate `dotenvy::var`, `dotenvy::from_filename*`